+++ inst/doc/char-part.R
@@ -1,11 +1,6 @@
 ## ---- setup, echo=FALSE-------------------------------------------------------
 knitr::opts_chunk$set(echo = TRUE, warning = FALSE, collapse = TRUE)
 
-## ---- eval = FALSE------------------------------------------------------------
-#  install.packages("EvoPhylo")
-#  ### OR
-#  devtools::install_github("tiago-simoes/EvoPhylo")
-
 ## -----------------------------------------------------------------------------
 library(EvoPhylo)
 
@@ -31,8 +26,11 @@ clusters <- make_clusters(dist_matrix, k = 3)
 plot(clusters)
 
 ## ---- eval = FALSE------------------------------------------------------------
-#  ## Write clusters to Nexus file
-#  cluster_to_nexus(clusters, file = "Clusters_Nexus.txt")
+#  ## Write clusters to Nexus file for Mr. Bayes
+#  cluster_to_nexus(clusters, file = "Clusters_MB.txt")
+#  
+#  ## Write partitioned alignments to separate Nexus files for BEAUTi
+#  write_partitioned_alignments(characters, clusters, file = "Clusters_BEAUTi.nex")
 
 ## ---- fig.width=10, fig.height=7, fig.align = "center", out.width = "100%"----
 #User may also generate clusters with PAM and produce a graphic clustering (tSNEs)
@@ -41,6 +39,9 @@ clusters <- make_clusters(dist_matrix, k = 3, tsne = TRUE, tsne_dim = 3)
 plot(clusters, nrow = 2, max.overlaps = 5)
 
 ## ---- eval = FALSE------------------------------------------------------------
-#  #Write clusters/partitions in Nexus file format
-#  cluster_to_nexus(clusters, file = "Clusters_Nexus.txt")
+#  ## Write clusters to Nexus file for Mr. Bayes
+#  cluster_to_nexus(clusters, file = "Clusters_MB.txt")
+#  
+#  ## Write partitioned alignments to separate Nexus files for BEAUTi
+#  write_partitioned_alignments(characters, clusters, file = "Clusters_BEAUTi.nex")